make '/unlock' route act as a drawer
