+++ src/app/(home)/page.tsx
@@ -1,0 +1,6 @@
+// The homepage content is rendered by the group layout (`HomeView`). This route
+// itself renders nothing — navigating to `/unlock` swaps in the unlock page
+// beside the same, still-mounted `HomeView`.
+export default function Home() {
+  return null;
+}
