+++ src/app/components/HomeView/index.ts
@@ -1,0 +1,1 @@
+export * from './HomeView';
