+++ src/app/components/index.ts
@@ -1,2 +1,3 @@
 export * from './HeroComponent';
+export * from './HomeView';
 export * from './HowToConnectComponent';
